Add new store movie detail
