+++ src/stores/movieDetail.ts
@@ -1,0 +1,13 @@
+import { ref } from 'vue'
+import { defineStore } from 'pinia'
+import type { Movie } from './types'
+
+export const useMovieDetailStore = defineStore('movieDetail', () => {
+  const movie = ref<Movie | null>(null)
+
+  function setMovie(value: Movie | null) {
+    movie.value = value
+  }
+
+  return { movie, setMovie }
+})
