+++ src/stores/types.ts
@@ -1,0 +1,9 @@
+export type Movie = {
+  title: string
+  year: string
+  genres: string[]
+  posterurl: string
+  imdbRating: number
+  duration: string
+  storyline: string
+}
